% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{dataModal}
\alias{dataModal}
\title{dbselect}
\usage{
dataModal(failed = FALSE)
}
\value{
data.table with average of values by frequenceSelected
}
\description{
Create dialog windows to download data in mod_access
}
