% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{\%!in\%}
\alias{\%!in\%}
\title{'%!in%'}
\usage{
x \%!in\% y
}
\arguments{
\item{dataSelected}{data.table from mod_accessdata}
}
\value{
return opposite of '%in%'
}
\description{
Select opposite of '%in%'
}
