% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesGraph.R
\name{graphWindRose}
\alias{graphWindRose}
\title{graphWindRose}
\usage{
graphWindRose(subsetoutbd)
}
\arguments{
\item{subsetoutbd}{data.table from mod_accessdata}
}
\value{
chart of Wind rose
}
\description{
Wind rose chart based on plot.windrose
}
