% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesCSS.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{grid_arrange_shared_legend}
\usage{
grid_arrange_shared_legend(
  ...,
  nrow = 1,
  ncol = length(list(...)),
  position = c("bottom", "right")
)
}
\value{
ggplots
}
\description{
Combined ggplots with a common legend
}
