% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesGraph.R
\name{plot.windrose}
\alias{plot.windrose}
\title{plot.windrose}
\source{
https://stackoverflow.com/questions/17266780/wind-rose-with-ggplot-r
}
\usage{
\method{plot}{windrose}(
  data,
  spd,
  dir,
  spdres = 2,
  dirres = 30,
  spdmin = 2,
  spdmax = 20,
  spdseq = NULL,
  palette = "YlGnBu",
  countmax = NA,
  debug = 0,
  titreLegend = "",
  titre = ""
)
}
\arguments{
\item{data}{data.frame}

\item{spd}{String of wind speed column}

\item{dir}{String of wind direction column}

\item{spdres}{Speed resolution}

\item{dirres}{Direction resolution}

\item{spdmin}{Speed scale min}

\item{spdmax}{Speed scale max}

\item{spdseq}{Speed scale sequence}

\item{palette}{RColorBrewer code palette (ex. "YlGnBu")}

\item{countmax}{NA by default}

\item{debug}{debug mode with 1, not with 0 (default)}

\item{titreLegend}{Title of legend}

\item{titre}{Title of chart}
}
\value{
chart of Wind rose
}
\description{
ggplot2 chart of wind rose
}
