% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{sqlOutputDatasetArchive}
\alias{sqlOutputDatasetArchive}
\title{sqlOutputDatasetArchive}
\usage{
sqlOutputDatasetArchive(language, checkinJeu, archiveType = "ZENODO")
}
\arguments{
\item{language}{String of language ("en" or "fr")}

\item{checkinJeu}{String of code_jeu (ex. "lgt-ges-eddycovariance")}

\item{archiveType}{String "OZCAR" to create archive type for SI Theia/OZCAR. If not, for ZENODO archive}
}
\value{
data.table of a dataset
}
\description{
Extract dataset by code_jeu attribute from data baseeCreate a file with header
}
