% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{writeHeaderFile}
\alias{writeHeaderFile}
\title{writeHeaderFile}
\usage{
writeHeaderFile(x, file, header, f = write.csv, ...)
}
\arguments{
\item{x}{data.frame or data.table}

\item{file}{String of file name}

\item{header}{String of header}

\item{f}{to write the file}

\item{...}{option for write.table}
}
\value{
Write a file
}
\description{
Create a file with header
}
