% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependanciesData.R
\name{writeHeaderPivot}
\alias{writeHeaderPivot}
\title{writeHeaderPivot}
\usage{
writeHeaderPivot(x, file, header, f = write.csv, ...)
}
\value{
Write a file in Pivot/TheiaOzcar format
}
\description{
Create a file in Pivot/TheiaOzcar format
}
